+++ pkg/analysis_server_client/CHANGELOG.md
@@ -1,3 +1,9 @@
+#1.1.0
+* Add analysis server protocol consts and classes
+* Overhaul the Server class
+* Add an example showing analysis of *.dart files in a directory
+* Update the required SDK
+
 #1.0.1
 * Update path on homepage